backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
langchain-openai
langgraph
langchain
python-dotenv
motor
pymongo
certifi
